Made changes to be compliant with @login_required relocation.
