+++ app/analytics/routes.py
@@ -5,7 +5,7 @@ from flask import render_template
 from flask_login import current_user
 
 from app.analytics import bp
-from app.manager.helpers import login_required
+from app.manager.permissions.utils import login_required
 from app.manager.db.db_interrogations import (
     query_expense_entries, query_revenue_entries, query_savings_entries, query_utilities_entries,
     get_expense_count, get_revenue_count, get_savings_count, get_utilities_count,
