+++ app/manager/permissions/routes.py
@@ -1,23 +1,13 @@
-from collections import defaultdict
-from typing import Iterator
-
 from flask import (
-    redirect, render_template, request, url_for, current_app
+    redirect, render_template, url_for, current_app
 )
 from flask_login import current_user
-from app.manager.helpers import login_required, extract_nested_strings
+from app.manager.permissions.utils import login_required, requires_permissions
 from app.manager.permissions import bp
 from app.manager.permissions.utils import convert_url_map_to_module_map
+from app.manager.db.db_interrogations import get_user_role_rules
 
 MODULE_EXCEPTIONS = ['debugtoolbar', 'api', 'auth', '_debug_toolbar', 'static', 'static_from_root']
-
-
-@bp.route('/user-roles/view', methods=('GET',))
-@login_required
-def user_roles_view():
-    user_id = current_user.get_id()
-
-    return render_template('manager/permissions/view_user_roles.html')
 
 
 # TODO Must have in mind the following: assuming that we have a collection of modules and submodules that upon user
@@ -26,10 +16,18 @@ def user_roles_view():
 #   If an endpoint is removed it is pretty easy, I suppose - you reparse the rules, create the JSON and update the DB
 #   but eliminate the differences between them.
 #   What if an endpoint is just plain modified?
-@bp.route('/user-roles/create', methods=('POST', 'GET'))
-def user_roles_create():
+@bp.route('/user-roles', methods=('GET',))
+@requires_permissions
+@login_required
+def user_roles():
+
     rule_map = current_app.url_map.iter_rules()
 
     module_map = convert_url_map_to_module_map(rule_map, skip_list=MODULE_EXCEPTIONS)
 
-    return render_template('manager/permissions/create_user_roles.html', module_map=module_map)
+    print(module_map)
+
+    return render_template('manager/permissions/user_roles.html', module_map=module_map)
+
+
+
